gemm workspace: 64-bit byte offsets — int32 overflow at many experts/rank
Both grouped-GEMM workspace kernels computed weight/output byte offsets in
int32: gather_rs (eid*N + sid*new_N)*K*elem and ag_scatter eid*N*K*elem
overflow at eid > 2^31/(N*K*2), yielding garbage weight pointers and WRONG
OUTPUT (no crash). K3 N=3584/K=3072 puts the cliff at ~97 experts/rank —
epr-bisected to exactly that boundary (96 pass / 104 fail); every
historical shape ran <=16 experts/rank and never hit it. Root cause of the
K3 l01/l1 allclose failures at small world sizes (W=4: 224 experts/rank);
4n16r K3 (56/rank) was unaffected, 2n (112/rank) would not have been.
M_acc*N terms widened too (latent at large-M cells). Post-fix: the failing
configs pass 4/4, N=5120/4096 regressions unchanged, 1n K3 l01 smoke 3/3.

+++ src/moe_ag_scatter/workspace_util.cu
@@ -340,9 +340,13 @@ prepare_workspace_kernel(
 
     problem_sizes[i] = cutlass::gemm::GemmCoord(Mi, N, K);
     ptr_A[i] = args.input;
-    ptr_B[i] = ptr_with_offset(args.weight[gid], eid * N * K * input_elem_size);
+    // 64-bit byte offsets (2026-08-21): int32 products overflow at many
+    // experts per rank — eid*N*K*2 crosses 2^31 at eid > 2^31/(N*K*2)
+    // (K3 ffn/H: eid ~97; the gather_rs sibling was the proven K3 layer1
+    // wrong-output cliff, epr-bisected to exactly that boundary).
+    ptr_B[i] = ptr_with_offset(args.weight[gid], (int64_t)eid * N * K * input_elem_size);
     ptr_C[i] = nullptr;
-    ptr_D[i] = ptr_with_offset(args.output[gid], M_acc * N * output_elem_size);
+    ptr_D[i] = ptr_with_offset(args.output[gid], (int64_t)M_acc * N * output_elem_size);
     lda[i] = GemmLayoutA::packed({Mi, K}).stride(0);
     ldb[i] = GemmLayoutB::packed({K, N}).stride(0);
     ldc[i] = GemmLayoutC::packed({Mi, N}).stride(0);
